Persist share/kocom/kocom.py in forked addon

+++ run.sh
@@ -2,11 +2,15 @@
 
 SHARE_DIR=/share/kocom
 
+mkdir -p $SHARE_DIR
+
 if [ ! -f $SHARE_DIR/kocom.conf ]; then
-	mkdir $SHARE_DIR
 	mv /kocom.conf $SHARE_DIR
 fi
-mv /kocom.py $SHARE_DIR
+
+if [ ! -f $SHARE_DIR/kocom.py ]; then
+	mv /kocom.py $SHARE_DIR
+fi
 
 echo "[Info] Run Kocom Wallpad with RS485!"
 cd $SHARE_DIR
